chore(build): pwa-wrapper as github dependency
Instead of using a local dependency definition for `pwa-wrapper`,
a github URL is used. This enables compiling and running the
application on other machines than the developer machine.

Using a local dependency path is only viable for local development
as long as the dependency itself is under constant development.
IMHO such dependencies should not be published but kept as
local change only.